move buttons from scheduling to common folder
update refs in scheduling comonents too
