+++ src/components/common/buttons/delete.vue
