+++ src/components/scheduling/magazines/list-item.vue
@@ -48,8 +48,8 @@
 <script>
 import mutation from '../../../graphql/scheduling/mutations/delete-magazine-schedule';
 import EditSchedule from './edit.vue';
-import EditButton from '../buttons/edit.vue';
-import DeleteButton from '../buttons/delete.vue';
+import EditButton from '../../common/buttons/edit.vue';
+import DeleteButton from '../../common/buttons/delete.vue';
 import OperationError from '../../operation-error.vue';
 
 export default {
