+++ src/components/scheduling/newsletters/list-item.vue
@@ -49,8 +49,8 @@
 import mutation from '../../../graphql/scheduling/mutations/delete-email-schedule';
 import DisplayDate from '../../display-date.vue';
 import EditSchedule from './edit.vue';
-import EditButton from '../buttons/edit.vue';
-import DeleteButton from '../buttons/delete.vue';
+import EditButton from '../../common/buttons/edit.vue';
+import DeleteButton from '../../common/buttons/delete.vue';
 import OperationError from '../../operation-error.vue';
 
 export default {
